feat: Add Forge-1.15.1 works, but some methods are SRG names
Forge for 1.15.1 is still in beta, so this shouldn't be published yet
and the branch cannot be merged to master until the function names
are changed in the mappings. I've updated two using the MCPBot, but
there was an issue with BlockPos.down() I didn't know how to handle.

See: http://export.mcpbot.bspk.rs/