feat: Forge 1.12 working mostly
Setup of the gradle build as a original style CoreMod with the
Mixin framework.  Seems source wall breaks are broken, but edge
erosion is working.

First pass requires significant changes to the common library files. I
will work this to find a better "interface cut" before merging to the
master branch.

An additional translationg layer to reduce the difference between the
1.12 code and the 1.14 code. The Block State object is named IBlockState
in 1.12 and just BlockState in 1.14+.

For: https://github.com/13rac1/Minecraft-Water-Erosion/issues/8