fix: forge-1.12: Fix DupeMod crash and include all files in shadowJar
FMLCorePluginContainsFMLMod was not set true in the Manifest, which causes
the server (only) to crash with:

Encountered an unexpected exception
net.minecraftforge.fml.common.DuplicateModsFoundException:
Duplicate Mods:
	water-erosion : /data/./mods/water-erosion-1.12.2-forge-1.1.0-alpha1-full.jar
	water-erosion : /data/./mods/water-erosion-1.12.2-forge-1.1.0-alpha1-full.jar

Setting it true makes the release Jar work on both server and clients, but
breaks the mod in development. Workaround is to set it in the Manifest during
the shadowJar build.

Now includes all needed files correctly during shadowJar build, specifically
including the mixin refmap and embedded Spongepowered Mixin classes.

Fixes #15
https://github.com/13rac1/Minecraft-Water-Erosion/issues/15